requirements.txt
matplotlib==3.5.2
inflect==5.6.0
librosa==0.9.1
scipy==1.8.1
Unidecode==1.3.4
pillow==8.4.0
pyworld==0.3.0
tqdm==4.64.0
g2p_en==2.1.0
wandb==0.12.18
streamlit==1.10.0
Cython==0.29.30
praat-parselmouth==0.4.1
lightning-flash==0.5.0
pytorch_lightning==1.4.0
transformers==4.1.1
protobuf==3.19.4
clean-text[gpl]==0.6.0
PyWavelets==1.3.0
omegaconf==2.2.2
text_unidecode
tgt==1.4.4
